webapi/openai_test.go
package main

import (
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestParseOpenAIMessages(t *testing.T) {
	t.Run("valid with image", func(t *testing.T) {
		messages := []OpenAIMessage{
			{
				Role:    "system",
				Content: "You are Sydney.",
			},
			{
				Role:    "user",
				Content: "Hello!",
			},
			{
				Role:    "assistant",
				Content: "Hi!",
			},
			{
				Role: "user",
				Content: []map[string]interface{}{
					{
						"type": "text",
						"text": "How are you?",
					},
					{
						"type": "image_url",
						"image_url": map[string]string{
							"url": "https://example.com/image.jpg",
						},
					},
				},
			},
		}
		result, err := ParseOpenAIMessages(messages)
		assert.Nil(t, err)
		assert.Equal(t, OpenAIMessagesParseResult{
			Prompt:         "How are you?",
			WebpageContext: "\n\n[system](#additional_instructions)\nYou are Sydney.\n\n[user](#message)\nHello!\n\n[assistant](#message)\nHi!",
			ImageURL:       "https://example.com/image.jpg",
		}, result)
	})
}
